refactor: Split ExerciceController logic into new controllers
